+++ lib/core/constants/hive_box_constants.dart
@@ -1,0 +1,5 @@
+class HiveBoxConstants {
+
+  static const String saveUser = 'saveUser';
+  
+}