+++ lib/core/services/device_id/device_id_service.dart
@@ -1,0 +1,5 @@
+abstract interface class DeviceIdService{
+
+  Future<String> getDeviceId();
+
+}