+++ lib/core/services/messages/messages.dart
@@ -1,0 +1,2 @@
+export 'message_service_impl.dart';
+export 'message_service.dart';