+++ lib/features/auth/domain/entities/entities.dart
@@ -1,0 +1,2 @@
+export 'role_entity.dart';
+export 'user_entity.dart';