+++ lib/features/auth/domain/entities/role_entity.dart
@@ -1,0 +1,15 @@
+import 'package:hive/hive.dart';
+
+part 'role_entity.g.dart';
+
+@HiveType(typeId: 1)
+class RoleEntity {
+
+  @HiveField(0)
+  final String name;
+
+  RoleEntity({
+    required this.name
+  });
+
+}